Created new TwoColorScatter component that takes in an additional set of shap values to create a color scale from blue to red on the graph. This graph comes with a legend that also acts as an adjustable slider to filter out the desired points.
I noticed Ani already setup the TGraph annotation and logic for building graphs, so I only fixed a bug with the brush in the Bar plot (since the brush worked even though an "annotation" (highlightedBars) was being passed in).

I added the ability (and connected with Ani's previous code) to pass in labels to be bolded for both the Swarm and Heatmap plots.

I added info about the SHAP Value/contribution to the top Paper component.